# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eko CP",
        "description": "Dokumentasi API untuk data Fase, Mapel, dan CP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/refleksi/kuis/{id}/{device_id}": {
            "get": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Get detail of specific kuis for a device",
                "description": "Mengambil detail kuis beserta status pengerjaan oleh device tertentu.",
                "operationId": "c48a011f817b56446384a8f66b590def",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "device_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuis detail retrieved successfully."
                    }
                }
            }
        },
        "/api/refleksi/pertanyaan/{kuis_id}": {
            "get": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Get next pertanyaan for a kuis",
                "operationId": "c0023db007c79512228c23766594eb6d",
                "parameters": [
                    {
                        "name": "kuis_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pertanyaan berhasil diambil."
                    }
                }
            }
        },
        "/api/refleksi/kuis/simpan": {
            "post": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Simpan hasil kuis",
                "operationId": "3ead2290d2c9ef593f32544873876947",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "kuis_id",
                                    "jawaban",
                                    "waktu_pengerjaan"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "string"
                                    },
                                    "kuis_id": {
                                        "type": "string"
                                    },
                                    "jawaban": {
                                        "type": "object"
                                    },
                                    "waktu_pengerjaan": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hasil kuis berhasil disimpan."
                    }
                }
            }
        },
...
        "/api/refleksi/harian/store": {
            "post": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Store new refleksi harian",
                "operationId": "e4c308a01b4cd9bc292c6e6bf207213f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "judul",
                                    "refleksi",
                                    "tanggal",
                                    "deskripsi"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "string"
                                    },
                                    "judul": {
                                        "type": "string"
                                    },
                                    "refleksi": {
                                        "type": "string"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refleksi created successfully."
                    }
                }
            }
        },
        "/api/refleksi/chat/{deviceId}": {
            "get": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Get chat history by device ID",
                "operationId": "4d3d526591e70bb4b08dc9e17367258b",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat history retrieved successfully."
                    }
                }
            }
        },
        "/api/refleksi/chat/send": {
            "post": {
                "tags": [
                    "Eko Refleksi"
                ],
                "summary": "Send a message to Groq chatbot",
                "operationId": "82e3ccee49bf715730466841bb471e01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "message"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message processed successfully."
                    }
                }
            }
        },
...
    },
...
}